requirements.txt
twisted==22.4.0
beautifulsoup4==4.12.3
itemadapter==0.3.0
psycopg2-binary==2.9.9
Requests==2.32.3
Scrapy==2.8.0
scrapy_fake_useragent==1.4.4
python-dotenv~=1.0.0
lxml>=4.9.0
APScheduler==3.11.0